refactor: extract template I/O into TemplateIO module (#1514)
Move Util::readTemplates / writeTemplates / getFolderTemplate into a
dedicated TemplateIO class (src/templateio.{h,cpp}), completing the Util
grab-bag split. Template handling (the .templates INI file and per-folder
.default_template lookup) is a self-contained concern used only by the
add-password flow in MainWindow.

- Verbatim move of all three impls.
- Update the two MainWindow callers to TemplateIO::.
- Retarget the readTemplates*/writeTemplates*/getFolderTemplate* cases in
  tst_util to TemplateIO.
- Drop now-dead includes from util.cpp (QSaveFile, QTextStream,
  QStringConverter, algorithm).
- Wire templateio.cpp/.h into src/src.pro.
